% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedSCORE.R
\name{getMaxDist}
\alias{getMaxDist}
\title{find the maxinum distance from the convex hull formed by the chosen K vertices}
\usage{
getMaxDist(centers, vertex.ind)
}
\arguments{
\item{centers}{L-by-(K-1) center matrix}

\item{vertex.ind}{index of the \code{K} centers that forms the convex hull}
}
\value{
the maximum distance
}
\description{
find the maxinum distance from the convex hull formed by the chosen K vertices
}
