% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedSCORE.R
\name{getMembership}
\alias{getMembership}
\title{calculated the membership of each node given ratio matrix and community centers}
\usage{
getMembership(R, vertices, K, eig.values, eig.vectors)
}
\arguments{
\item{R}{n-by-(K-1) ratio matrix}

\item{vertices}{K-by-(K-1) community centers}

\item{K}{number of communities.}

\item{eig.values}{eigenvalues of adjacency matrix.}

\item{eig.vectors}{eigenvectors of adjacency matrix.}
}
\value{
n-by-K membership matrix
}
\description{
calculated the membership of each node given ratio matrix and community centers
}
