% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedSCORE.R
\name{vertexHunting}
\alias{vertexHunting}
\title{Vertex hunting algorithm to find the cluster centers}
\usage{
vertexHunting(R, K, verbose = F)
}
\arguments{
\item{R}{n-by-(K-1) ratio matrix}

\item{K}{number of communities.}

\item{verbose}{whether or not to show a progress bar}
}
\description{
Vertex hunting algorithm to find the cluster centers
}
