% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedSCORE.R
\name{vertexSearch}
\alias{vertexSearch}
\title{select the \code{K} vertices from given \code{L} centers}
\usage{
vertexSearch(centers, K)
}
\arguments{
\item{centers}{L-by-(K-1) center matrix}

\item{K}{number of communities.}
}
\value{
A list containing \describe{
  \item{ind}{a vector of \code{K} integers indicating the index of selected \code{K} vertices out of \code{L} centers.}
  \item{dist}{The maximum distance from centers to the convex hull formed by the \code{K} selected vertice}
}
}
\description{
select the \code{K} vertices from given \code{L} centers
}
